refactor(tui): refactor album, song, and artist views into generic ItemView
