+++ tui/src/ui/components/content_view/views/artist.rs
@@ -3,20 +3,19 @@
 use std::{fmt::Display, sync::Mutex};
 
 use crossterm::event::{KeyCode, KeyEvent, MouseEvent};
-use mecomp_core::format_duration;
 use mecomp_storage::db::schemas::artist::Artist;
 use ratatui::{
-    layout::{Alignment, Constraint, Direction, Layout, Margin, Rect},
+    layout::{Margin, Rect},
     style::{Style, Stylize},
     text::{Line, Span},
-    widgets::{Block, Borders, Paragraph, Scrollbar, ScrollbarOrientation},
+    widgets::{Block, Borders, Scrollbar, ScrollbarOrientation},
 };
 use tokio::sync::mpsc::UnboundedSender;
 
 use crate::{
     state::action::{Action, AudioAction, PopupAction, QueueAction},
     ui::{
-        colors::{BORDER_FOCUSED, BORDER_UNFOCUSED, TEXT_HIGHLIGHT, TEXT_NORMAL},
+        colors::{BORDER_FOCUSED, BORDER_UNFOCUSED, TEXT_HIGHLIGHT},
         components::{content_view::ActiveView, Component, ComponentRender, RenderProps},
         widgets::{
             popups::PopupType,
@@ -27,270 +26,11 @@ use crate::{
 };
 
 use super::{
-    checktree_utils::{
-        construct_add_to_playlist_action, construct_add_to_queue_action,
-        construct_start_radio_action, create_album_tree_item, create_artist_tree_leaf,
-        create_song_tree_item,
-    },
-    ArtistViewProps, RADIO_SIZE,
+    checktree_utils::create_artist_tree_leaf, generic::ItemView, ArtistViewProps, RADIO_SIZE,
 };
 
 #[allow(clippy::module_name_repetitions)]
-pub struct ArtistView {
-    /// Action Sender
-    pub action_tx: UnboundedSender<Action>,
-    /// Mapped Props from state
-    pub props: Option<ArtistViewProps>,
-    /// tree state
-    tree_state: Mutex<CheckTreeState<String>>,
-}
-
-impl Component for ArtistView {
-    fn new(state: &AppState, action_tx: UnboundedSender<Action>) -> Self
-    where
-        Self: Sized,
-    {
-        Self {
-            action_tx,
-            props: state.additional_view_data.artist.clone(),
-            tree_state: Mutex::new(CheckTreeState::default()),
-        }
-    }
-
-    fn move_with_state(self, state: &AppState) -> Self
-    where
-        Self: Sized,
-    {
-        if let Some(props) = &state.additional_view_data.artist {
-            Self {
-                props: Some(props.to_owned()),
-                tree_state: Mutex::new(CheckTreeState::default()),
-                ..self
-            }
-        } else {
-            self
-        }
-    }
-
-    fn name(&self) -> &str {
-        "Artist View"
-    }
-
-    fn handle_key_event(&mut self, key: KeyEvent) {
-        match key.code {
-            // arrow keys
-            KeyCode::Up => {
-                self.tree_state.lock().unwrap().key_up();
-            }
-            KeyCode::Down => {
-                self.tree_state.lock().unwrap().key_down();
-            }
-            KeyCode::Left => {
-                self.tree_state.lock().unwrap().key_left();
-            }
-            KeyCode::Right => {
-                self.tree_state.lock().unwrap().key_right();
-            }
-            KeyCode::Char(' ') => {
-                self.tree_state.lock().unwrap().key_space();
-            }
-            // Enter key opens selected view
-            KeyCode::Enter => {
-                if self.tree_state.lock().unwrap().toggle_selected() {
-                    let things = self.tree_state.lock().unwrap().get_selected_thing();
-
-                    if let Some(thing) = things {
-                        self.action_tx
-                            .send(Action::SetCurrentView(thing.into()))
-                            .unwrap();
-                    }
-                }
-            }
-            // if there are checked items, add them to the queue, otherwise send the whole artist to the queue
-            KeyCode::Char('q') => {
-                let checked_things = self.tree_state.lock().unwrap().get_checked_things();
-                if let Some(action) = construct_add_to_queue_action(
-                    checked_things,
-                    self.props.as_ref().map(|p| &p.id),
-                ) {
-                    self.action_tx.send(action).unwrap();
-                }
-            }
-            // if there are checked items, start radio from checked items, otherwise start radio from artist
-            KeyCode::Char('r') => {
-                let checked_things = self.tree_state.lock().unwrap().get_checked_things();
-                if let Some(action) =
-                    construct_start_radio_action(checked_things, self.props.as_ref().map(|p| &p.id))
-                {
-                    self.action_tx.send(action).unwrap();
-                }
-            }
-            // if there are checked items, add them to playlist, otherwise add the whole artist to playlist
-            KeyCode::Char('p') => {
-                let checked_things = self.tree_state.lock().unwrap().get_checked_things();
-                if let Some(action) = construct_add_to_playlist_action(
-                    checked_things,
-                    self.props.as_ref().map(|p| &p.id),
-                ) {
-                    self.action_tx.send(action).unwrap();
-                }
-            }
-            _ => {}
-        }
-    }
-
-    fn handle_mouse_event(&mut self, mouse: MouseEvent, area: Rect) {
-        // adjust the area to account for the border
-        let area = area.inner(Margin::new(1, 1));
-        let [_, content_area] = split_area(area);
-        let content_area = Rect {
-            y: content_area.y + 2,
-            height: content_area.height - 2,
-            ..content_area
-        };
-
-        let result = self
-            .tree_state
-            .lock()
-            .unwrap()
-            .handle_mouse_event(mouse, content_area);
-        if let Some(action) = result {
-            self.action_tx.send(action).unwrap();
-        }
-    }
-}
-
-fn split_area(area: Rect) -> [Rect; 2] {
-    let [info_area, content_area] = *Layout::default()
-        .direction(Direction::Vertical)
-        .constraints([Constraint::Length(3), Constraint::Min(4)])
-        .split(area)
-    else {
-        panic!("Failed to split artist view area")
-    };
-
-    [info_area, content_area]
-}
-
-impl ComponentRender<RenderProps> for ArtistView {
-    fn render_border(&self, frame: &mut ratatui::Frame, props: RenderProps) -> RenderProps {
-        let border_style = if props.is_focused {
-            Style::default().fg(BORDER_FOCUSED.into())
-        } else {
-            Style::default().fg(BORDER_UNFOCUSED.into())
-        };
-
-        // draw borders and get content area
-        let area = if let Some(state) = &self.props {
-            let border = Block::bordered()
-                .title_top("Artist View")
-                .title_bottom(" \u{23CE} : Open | ←/↑/↓/→: Navigate | \u{2423} Check")
-                .border_style(border_style);
-            let content_area = border.inner(props.area);
-            frame.render_widget(border, props.area);
-
-            // split the area to make room for the artist info
-            let [info_area, content_area] = split_area(content_area);
-
-            // render the artist info
-            frame.render_widget(
-                Paragraph::new(vec![
-                    Line::from(Span::styled(
-                        state.artist.name.to_string(),
-                        Style::default().bold(),
-                    )),
-                    Line::from(vec![
-                        Span::raw("Albums: "),
-                        Span::styled(
-                            state.artist.album_count.to_string(),
-                            Style::default().italic(),
-                        ),
-                        Span::raw("  Songs: "),
-                        Span::styled(
-                            state.artist.song_count.to_string(),
-                            Style::default().italic(),
-                        ),
-                        Span::raw("  Duration: "),
-                        Span::styled(
-                            format_duration(&state.artist.runtime),
-                            Style::default().italic(),
-                        ),
-                    ]),
-                ])
-                .alignment(Alignment::Center),
-                info_area,
-            );
-
-            // draw an additional border around the content area to display additional instructions
-            let border = Block::default()
-                .borders(Borders::TOP)
-                .title_top("q: add to queue | r: start radio | p: add to playlist")
-                .border_style(border_style);
-            frame.render_widget(&border, content_area);
-            let content_area = border.inner(content_area);
-
-            // draw an additional border around the content area to indicate whether operations will be performed on the entire item, or just the checked items
-            let border = Block::default()
-                .borders(Borders::TOP)
-                .title_top(Line::from(vec![
-                    Span::raw("Performing operations on "),
-                    Span::raw(
-                        if self
-                            .tree_state
-                            .lock()
-                            .unwrap()
-                            .get_checked_things()
-                            .is_empty()
-                        {
-                            "entire artist"
-                        } else {
-                            "checked items"
-                        },
-                    )
-                    .fg(TEXT_HIGHLIGHT),
-                ]))
-                .italic()
-                .border_style(border_style);
-            frame.render_widget(&border, content_area);
-            border.inner(content_area)
-        } else {
-            let border = Block::bordered()
-                .title_top("Artist View")
-                .border_style(border_style);
-            frame.render_widget(&border, props.area);
-            border.inner(props.area)
-        };
-
-        RenderProps { area, ..props }
-    }
-
-    fn render_content(&self, frame: &mut ratatui::Frame, props: RenderProps) {
-        if let Some(state) = &self.props {
-            // create list to hold artist albums and songs
-            let album_tree = create_album_tree_item(&state.albums).unwrap();
-            let song_tree = create_song_tree_item(&state.songs).unwrap();
-            let items = &[album_tree, song_tree];
-
-            // render the artist artists / album
-            frame.render_stateful_widget(
-                CheckTree::new(items)
-                    .unwrap()
-                    .highlight_style(Style::default().fg(TEXT_HIGHLIGHT.into()).bold()),
-                props.area,
-                &mut self.tree_state.lock().unwrap(),
-            );
-        } else {
-            let text = "No active artist";
-
-            frame.render_widget(
-                Line::from(text)
-                    .style(Style::default().fg(TEXT_NORMAL.into()))
-                    .alignment(Alignment::Center),
-                props.area,
-            );
-        }
-    }
-}
+pub type ArtistView = ItemView<ArtistViewProps>;
 
 pub struct LibraryArtistsView {
     /// Action Sender
