+++ tui/src/ui/components/content_view/views/song.rs
@@ -5,17 +5,17 @@ use std::{fmt::Display, sync::Mutex};
 use crossterm::event::{KeyCode, KeyEvent, MouseEvent};
 use mecomp_storage::db::schemas::song::Song;
 use ratatui::{
-    layout::{Alignment, Constraint, Direction, Layout, Margin, Rect},
+    layout::{Margin, Rect},
     style::{Style, Stylize},
     text::{Line, Span},
-    widgets::{Block, Borders, Paragraph, Scrollbar, ScrollbarOrientation},
+    widgets::{Block, Borders, Scrollbar, ScrollbarOrientation},
 };
 use tokio::sync::mpsc::UnboundedSender;
 
 use crate::{
     state::action::{Action, AudioAction, PopupAction, QueueAction},
     ui::{
-        colors::{BORDER_FOCUSED, BORDER_UNFOCUSED, TEXT_HIGHLIGHT, TEXT_NORMAL},
+        colors::{BORDER_FOCUSED, BORDER_UNFOCUSED, TEXT_HIGHLIGHT},
         components::{content_view::ActiveView, Component, ComponentRender, RenderProps},
         widgets::{
             popups::PopupType,
@@ -25,293 +25,10 @@ use crate::{
     },
 };
 
-use super::{
-    checktree_utils::{
-        construct_add_to_playlist_action, construct_add_to_queue_action,
-        construct_start_radio_action, create_album_tree_leaf, create_artist_tree_item,
-        create_collection_tree_item, create_playlist_tree_item, create_song_tree_leaf,
-    },
-    SongViewProps, RADIO_SIZE,
-};
+use super::{checktree_utils::create_song_tree_leaf, generic::ItemView, SongViewProps, RADIO_SIZE};
 
 #[allow(clippy::module_name_repetitions)]
-pub struct SongView {
-    /// Action Sender
-    pub action_tx: UnboundedSender<Action>,
-    /// Mapped Props from state
-    pub props: Option<SongViewProps>,
-    /// tree state
-    tree_state: Mutex<CheckTreeState<String>>,
-}
-
-impl Component for SongView {
-    fn new(state: &AppState, action_tx: UnboundedSender<Action>) -> Self
-    where
-        Self: Sized,
-    {
-        Self {
-            action_tx,
-            props: state.additional_view_data.song.clone(),
-            tree_state: Mutex::new(CheckTreeState::default()),
-        }
-    }
-
-    fn move_with_state(self, state: &AppState) -> Self
-    where
-        Self: Sized,
-    {
-        if let Some(props) = &state.additional_view_data.song {
-            Self {
-                props: Some(props.to_owned()),
-                tree_state: Mutex::new(CheckTreeState::default()),
-                ..self
-            }
-        } else {
-            self
-        }
-    }
-
-    fn name(&self) -> &str {
-        "Song View"
-    }
-
-    fn handle_key_event(&mut self, key: KeyEvent) {
-        match key.code {
-            // arrow keys
-            KeyCode::Up => {
-                self.tree_state.lock().unwrap().key_up();
-            }
-            KeyCode::Down => {
-                self.tree_state.lock().unwrap().key_down();
-            }
-            KeyCode::Left => {
-                self.tree_state.lock().unwrap().key_left();
-            }
-            KeyCode::Right => {
-                self.tree_state.lock().unwrap().key_right();
-            }
-            KeyCode::Char(' ') => {
-                self.tree_state.lock().unwrap().key_space();
-            }
-            // Enter key opens selected view
-            KeyCode::Enter => {
-                if self.tree_state.lock().unwrap().toggle_selected() {
-                    let things = self.tree_state.lock().unwrap().get_selected_thing();
-
-                    if let Some(thing) = things {
-                        self.action_tx
-                            .send(Action::SetCurrentView(thing.into()))
-                            .unwrap();
-                    }
-                }
-            }
-            // if there are checked items, add them to the queue, otherwise send the song to the queue
-            KeyCode::Char('q') => {
-                let checked_things = self.tree_state.lock().unwrap().get_checked_things();
-                if let Some(action) = construct_add_to_queue_action(
-                    checked_things,
-                    self.props.as_ref().map(|p| &p.id),
-                ) {
-                    self.action_tx.send(action).unwrap();
-                }
-            }
-            // if there are checked items, start radio from checked items, otherwise start radio from song
-            KeyCode::Char('r') => {
-                let checked_things = self.tree_state.lock().unwrap().get_checked_things();
-                if let Some(action) =
-                    construct_start_radio_action(checked_things, self.props.as_ref().map(|p| &p.id))
-                {
-                    self.action_tx.send(action).unwrap();
-                }
-            }
-            // if there are checked items, add them to playlist, otherwise add the song to playlist
-            KeyCode::Char('p') => {
-                let checked_things = self.tree_state.lock().unwrap().get_checked_things();
-                if let Some(action) = construct_add_to_playlist_action(
-                    checked_things,
-                    self.props.as_ref().map(|p| &p.id),
-                ) {
-                    self.action_tx.send(action).unwrap();
-                }
-            }
-            _ => {}
-        }
-    }
-
-    fn handle_mouse_event(&mut self, mouse: MouseEvent, area: Rect) {
-        // adjust the area to account for the border
-        let area = area.inner(Margin::new(1, 1));
-        let [_, content_area] = split_area(area);
-        let content_area = Rect {
-            y: content_area.y + 2,
-            height: content_area.height - 2,
-            ..content_area
-        };
-
-        let result = self
-            .tree_state
-            .lock()
-            .unwrap()
-            .handle_mouse_event(mouse, content_area);
-        if let Some(action) = result {
-            self.action_tx.send(action).unwrap();
-        }
-    }
-}
-
-fn split_area(area: Rect) -> [Rect; 2] {
-    let [info_area, content_area] = *Layout::default()
-        .direction(Direction::Vertical)
-        .constraints([Constraint::Length(3), Constraint::Min(4)])
-        .split(area)
-    else {
-        panic!("Failed to split song view area")
-    };
-
-    [info_area, content_area]
-}
-
-impl ComponentRender<RenderProps> for SongView {
-    fn render_border(&self, frame: &mut ratatui::Frame, props: RenderProps) -> RenderProps {
-        let border_style = if props.is_focused {
-            Style::default().fg(BORDER_FOCUSED.into())
-        } else {
-            Style::default().fg(BORDER_UNFOCUSED.into())
-        };
-
-        let border = Block::bordered()
-            .title_top("Song View")
-            .border_style(border_style);
-        frame.render_widget(&border, props.area);
-        // draw borders and get area for the content (album and artists of song)
-        let area = self.props.as_ref().map_or_else(
-            || border.inner(props.area),
-            |state| {
-                let area = border.inner(props.area);
-
-                // split area to make room for song info
-                let [info_area, content_area] = split_area(area);
-
-                // render the song info
-                frame.render_widget(
-                    Paragraph::new(vec![
-                        Line::from(vec![
-                            Span::styled(state.song.title.to_string(), Style::default().bold()),
-                            Span::raw(" "),
-                            Span::styled(
-                                state
-                                    .song
-                                    .artist
-                                    .iter()
-                                    .map(ToString::to_string)
-                                    .collect::<Vec<String>>()
-                                    .join(", "),
-                                Style::default().italic(),
-                            ),
-                        ]),
-                        Line::from(vec![
-                            Span::raw("Track/Disc: "),
-                            Span::styled(
-                                format!(
-                                    "{}/{}",
-                                    state.song.track.unwrap_or_default(),
-                                    state.song.disc.unwrap_or_default()
-                                ),
-                                Style::default().italic(),
-                            ),
-                            Span::raw("  Duration: "),
-                            Span::styled(
-                                format!(
-                                    "{}:{:04.1}",
-                                    state.song.runtime.as_secs() / 60,
-                                    state.song.runtime.as_secs_f32() % 60.0,
-                                ),
-                                Style::default().italic(),
-                            ),
-                            Span::raw("  Genre(s): "),
-                            Span::styled(
-                                state
-                                    .song
-                                    .genre
-                                    .iter()
-                                    .map(ToString::to_string)
-                                    .collect::<Vec<String>>()
-                                    .join(", "),
-                                Style::default().italic(),
-                            ),
-                        ]),
-                    ])
-                    .alignment(Alignment::Center),
-                    info_area,
-                );
-
-                // draw an additional border around the content area to display additional instructions
-                let border = Block::new()
-                    .borders(Borders::TOP)
-                    .title_top("q: add to queue | r: start radio | p: add to playlist")
-                    .border_style(border_style);
-                frame.render_widget(&border, content_area);
-                let content_area = border.inner(content_area);
-
-                // draw an additional border around the content area to indicate whether operations will be performed on the entire item, or just the checked items
-                let border = Block::default()
-                    .borders(Borders::TOP)
-                    .title_top(Line::from(vec![
-                        Span::raw("Performing operations on "),
-                        Span::raw(
-                            if self
-                                .tree_state
-                                .lock()
-                                .unwrap()
-                                .get_checked_things()
-                                .is_empty()
-                            {
-                                "the song"
-                            } else {
-                                "checked items"
-                            },
-                        )
-                        .fg(TEXT_HIGHLIGHT),
-                    ]))
-                    .italic()
-                    .border_style(border_style);
-                frame.render_widget(&border, content_area);
-                border.inner(content_area)
-            },
-        );
-
-        RenderProps { area, ..props }
-    }
-
-    fn render_content(&self, frame: &mut ratatui::Frame, props: RenderProps) {
-        if let Some(state) = &self.props {
-            // create a tree to hold song album and artists
-            let album_tree = create_album_tree_leaf(&state.album, Some(Span::raw("Album: ")));
-            let artist_tree = create_artist_tree_item(state.artists.as_slice()).unwrap();
-            let playlist_tree = create_playlist_tree_item(&state.playlists).unwrap();
-            let collection_tree = create_collection_tree_item(&state.collections).unwrap();
-            let items = &[artist_tree, album_tree, playlist_tree, collection_tree];
-
-            // render the song artists / album tree
-            frame.render_stateful_widget(
-                CheckTree::new(items)
-                    .unwrap()
-                    .highlight_style(Style::default().fg(TEXT_HIGHLIGHT.into()).bold()),
-                props.area,
-                &mut self.tree_state.lock().unwrap(),
-            );
-        } else {
-            let text = "No active song";
-
-            frame.render_widget(
-                Line::from(text)
-                    .style(Style::default().fg(TEXT_NORMAL.into()))
-                    .alignment(Alignment::Center),
-                props.area,
-            );
-        }
-    }
-}
+pub type SongView = ItemView<SongViewProps>;
 
 pub struct LibrarySongsView {
     /// Action Sender
@@ -807,7 +524,7 @@ mod item_view_tests {
             "│▶ Playlists (0):                                          │",
             "│▶ Collections (0):                                        │",
             "│                                                          │",
-            "└──────────────────────────────────────────────────────────┘",
+            "└ ⏎ : Open | ←/↑/↓/→: Navigate | ␣ Check───────────────────┘",
         ]);
 
         assert_buffer_eq(&buffer, &expected);
@@ -841,7 +558,7 @@ mod item_view_tests {
             "│▼ Playlists (0):                                          │",
             "│                                                          │",
             "│▶ Collections (0):                                        │",
-            "└──────────────────────────────────────────────────────────┘",
+            "└ ⏎ : Open | ←/↑/↓/→: Navigate | ␣ Check───────────────────┘",
         ]);
         assert_buffer_eq(&buffer, &expected);
 
@@ -863,7 +580,7 @@ mod item_view_tests {
             "│▶ Playlists (0):                                          │",
             "│▶ Collections (0):                                        │",
             "│                                                          │",
-            "└──────────────────────────────────────────────────────────┘",
+            "└ ⏎ : Open | ←/↑/↓/→: Navigate | ␣ Check───────────────────┘",
         ]);
         assert_buffer_eq(&buffer, &expected);
 
@@ -888,7 +605,7 @@ mod item_view_tests {
             "│▶ Playlists (0):                                          │",
             "│▼ Collections (0):                                        │",
             "│                                                          │",
-            "└──────────────────────────────────────────────────────────┘",
+            "└ ⏎ : Open | ←/↑/↓/→: Navigate | ␣ Check───────────────────┘",
         ]);
         assert_buffer_eq(&buffer, &expected);
 
@@ -922,7 +639,7 @@ mod item_view_tests {
             "│▶ Playlists (1):                                          │",
             "│▶ Collections (1):                                        │",
             "│                                                          │",
-            "└──────────────────────────────────────────────────────────┘",
+            "└ ⏎ : Open | ←/↑/↓/→: Navigate | ␣ Check───────────────────┘",
         ]);
 
         assert_buffer_eq(&buffer, &expected);
@@ -953,7 +670,7 @@ mod item_view_tests {
             "│Performing operations on the song─────────────────────────│",
             "│▶ Artists (1):                                            │",
             "│☐ Album: Test Album Test Artist                           │",
-            "└──────────────────────────────────────────────────────────┘",
+            "└ ⏎ : Open | ←/↑/↓/→: Navigate | ␣ Check───────────────────┘",
         ]);
         assert_buffer_eq(&buffer, &expected);
 
@@ -976,7 +693,7 @@ mod item_view_tests {
             "│Performing operations on checked items────────────────────│",
             "│▶ Artists (1):                                            │",
             "│☑ Album: Test Album Test Artist                           │",
-            "└──────────────────────────────────────────────────────────┘",
+            "└ ⏎ : Open | ←/↑/↓/→: Navigate | ␣ Check───────────────────┘",
         ]);
 
         assert_buffer_eq(&buffer, &expected);
@@ -1113,7 +830,7 @@ mod item_view_tests {
             "│Performing operations on the song─────────────────────────│",
             "│▶ Artists (1):                                            │",
             "│☐ Album: Test Album Test Artist                           │",
-            "└──────────────────────────────────────────────────────────┘",
+            "└ ⏎ : Open | ←/↑/↓/→: Navigate | ␣ Check───────────────────┘",
         ]);
         assert_buffer_eq(&buffer, &expected);
 
@@ -1141,7 +858,7 @@ mod item_view_tests {
             "│Performing operations on the song─────────────────────────│",
             "│▼ Artists (1):                                            │",
             "│  ☐ Test Artist                                           │",
-            "└──────────────────────────────────────────────────────────┘",
+            "└ ⏎ : Open | ←/↑/↓/→: Navigate | ␣ Check───────────────────┘",
         ]);
         assert_buffer_eq(&buffer, &expected);
 
@@ -1190,7 +907,7 @@ mod item_view_tests {
             "│Performing operations on checked items────────────────────│",
             "│▼ Artists (1):                                            │",
             "│  ☑ Test Artist                                           │",
-            "└──────────────────────────────────────────────────────────┘",
+            "└ ⏎ : Open | ←/↑/↓/→: Navigate | ␣ Check───────────────────┘",
         ]);
         assert_buffer_eq(&buffer, &expected);
